同步EUI至ET7.2Release版本，方便后续MMORPG-战斗篇课程制作

+++ Unity/Assets/Scripts/Codes/Hotfix/Share/Module/Scene/ClientSceneManagerComponentSystem.cs
@@ -26,13 +26,13 @@
             return ClientSceneManagerComponent.Instance.Get(entity.DomainZone());
         }
         
-        public static Scene Get(this ClientSceneManagerComponent self, long id)
+        public static Scene Get(this ClientSceneManagerComponent self, int id)
         {
             Scene scene = self.GetChild<Scene>(id);
             return scene;
         }
         
-        public static void Remove(this ClientSceneManagerComponent self, long id)
+        public static void Remove(this ClientSceneManagerComponent self, int id)
         {
             self.RemoveChild(id);
         }
